+++ orm/src/test/java/org/hibernate/tool/internal/reveng/strategy/MetaAttributeHelperTest.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.reveng;
+package org.hibernate.tool.internal.reveng.strategy;
 
 import java.io.ByteArrayInputStream;
 import java.util.Collection;
@@ -8,7 +8,8 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 
 import org.apache.commons.collections4.MultiValuedMap;
-import org.hibernate.tool.internal.reveng.MetaAttributeHelper.SimpleMetaAttribute;
+import org.hibernate.tool.internal.reveng.strategy.MetaAttributeHelper;
+import org.hibernate.tool.internal.reveng.strategy.MetaAttributeHelper.SimpleMetaAttribute;
 import org.junit.Assert;
 import org.junit.Test;
 import org.w3c.dom.Document;
